Add C++ solutions for various problems and update documentation

+++ src/cpp/0003_longest_substring_without_repeating_characters.cc
@@ -4,12 +4,12 @@
 using namespace std;
 
 int lengthOfLongestSubstring(string s) {
-    size_t n = s.length();
-    size_t res = 0;
-    size_t left = 0;
+    int n = s.length();
+    int res = 0;
+    int left = 0;
     unordered_set<char> window;
 
-    for (size_t right = 0; right < n; right++) {
+    for (int right = 0; right < n; right++) {
         char ch = s[right];
 
         while (window.find(ch) != window.end()) {
